# generated-openapi-client/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Trieve API",
    "description": "Trieve OpenAPI Specification. This document describes all of the operations available through the Trieve API.",
    "contact": {
      "name": "Trieve Team",
      "url": "https://trieve.ai",
      "email": "developers@trieve.ai"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/devflowinc/actix-web-template/blob/main/LICENSE.txt"
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:8090",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/contacts": {
      "post": {
        "tags": [
          "contacts"
        ],
        "operationId": "create_contact",
        "requestBody": {
          "description": "JSON request payload to create a new contact",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContactReqPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "JSON body representing the contact that was created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Org"
                }
              }
            }
          },
          "401": {
            "description": "Service error relating to authentication status of the user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespPayload"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [
              "readonly"
            ]
          }
        ]
      }
    },
    "/api/contacts/{contact_id}": {
      "get": {
        "tags": [
          "contacts"
        ],
        "operationId": "get_contact",
        "parameters": [
          {
            "name": "contact_id",
            "in": "path",
            "description": "The contacts id to use for the request",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JSON object representing the requested contact",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Org"
                }
              }
            }
          },
          "401": {
            "description": "Service error relating to authentication status of the user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespPayload"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [
              "readonly"
            ]
          }
        ]
      },
      "put": {
        "tags": [
          "contacts"
        ],
        "operationId": "update_contact",
        "parameters": [
          {
            "name": "contact_id",
            "in": "path",
            "description": "The contacts id to use for the request",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "JSON request payload to update the contact",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContactReqPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Object representing the renamed contact",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Org"
                }
              }
            }
          },
          "401": {
            "description": "Service error relating to authentication status of the user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespPayload"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [
              "readonly"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "contacts"
        ],
        "operationId": "delete_contact",
        "parameters": [
          {
            "name": "contact_id",
            "in": "path",
            "description": "The contacts id to use for the request",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No content response indicating that the contacts was successfully deleted"
          },
          "401": {
            "description": "Service error relating to authentication status of the user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespPayload"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [
              "readonly"
            ]
          }
        ]
      }
    },
    "/api/deals": {
      "post": {
        "tags": [
          "deals"
        ],
        "operationId": "create_deal",
        "requestBody": {
          "description": "JSON request payload to create a new deal",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDealReqPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "JSON body representing the deal that was created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Org"
                }
              }
            }
          },
          "401": {
            "description": "Service error relating to authentication status of the user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespPayload"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [
              "readonly"
            ]
          }
        ]
      }
    },
    "/api/deals/{deal_id}": {
      "get": {
        "tags": [
          "deals"
        ],
        "operationId": "get_deal",
        "parameters": [
          {
            "name": "deal_id",
            "in": "path",
            "description": "The deal id to use for the request",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JSON object representing the requested deal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Org"
                }
              }
            }
          },
          "401": {
            "description": "Service error relating to authentication status of the user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespPayload"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [
              "readonly"
            ]
          }
        ]
      },
      "put": {
        "tags": [
          "deals"
        ],
        "operationId": "update_deal",
        "parameters": [
          {
            "name": "deal_id",
            "in": "path",
            "description": "The deal id to use for the request",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "JSON request payload to update the deal",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDealReqPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Object representing the renamed deal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Org"
                }
              }
            }
          },
          "401": {
            "description": "Service error relating to authentication status of the user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespPayload"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [
              "readonly"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "deals"
        ],
        "operationId": "delete_deal",
        "parameters": [
          {
            "name": "deal_id",
            "in": "path",
            "description": "The deal id to use for the request",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No content response indicating that the deal was successfully deleted"
          },
          "401": {
            "description": "Service error relating to authentication status of the user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespPayload"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [
              "readonly"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateContactReqPayload": {
        "type": "object",
        "required": [
          "first_name",
          "last_name"
        ],
        "properties": {
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          }
        }
      },
      "CreateDealReqPayload": {
        "type": "object",
        "properties": {
          "active": {
            "type": "boolean",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "size": {
            "type": "number",
            "format": "float",
            "nullable": true
          }
        }
      },
...
      "ErrorRespPayload": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "example": {
          "message": "Bad Request"
        }
      },
...
      "Org": {
        "type": "object",
        "required": [
          "id",
          "name",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "$ref": "#/components/schemas/PrefixedUuid"
          },
          "name": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "example": {
          "created_at": "2021-01-01T00:00:00",
          "id": "e3e3e3e3-e3e3-e3e3-e3e3-e3e3e3e3e3e3",
          "name": "Trieve Team",
          "updated_at": "2021-01-01T00:00:00"
        }
      },
      "PrefixedUuid": {
        "type": "string"
      },
      "UpdateContactReqPayload": {
        "type": "object",
        "properties": {
          "first_name": {
            "type": "string",
            "nullable": true
          },
          "last_name": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "UpdateDealReqPayload": {
        "type": "object",
        "properties": {
          "active": {
            "type": "boolean",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "size": {
            "type": "number",
            "format": "float",
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    }
  },
...
}